requirements.txt
Django==1.3.4
MySQL-python==1.2.4c1
Pillow==1.7.8
South==0.7.6
Whoosh==2.4.1
amqp==1.0.6
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.19
celery==3.0.12
distribute==0.6.30
dj-database-url==0.2.1
django-celery==3.0.11
django-haystack==1.2.7
easy-thumbnails==1.1
kombu==2.5.4
psycopg2==2.4.5
python-dateutil==1.5
pytz==2012h
wsgiref==0.1.2
